requirements/pro.txt
django==2.1.2
pillow==5.3.0
mysqlclient==1.3.13
ipython==7.1.1
pyyaml==3.13
ruamel.yaml==0.15.80
python-nmap==0.6.1
redis==3.0.1
pymongo==3.7.1
paramiko==2.4.2
django-simple-history==2.6.0
celery==4.2.1
celery-once==2.0.0
flower==0.9.2
